Fix VerifyError: create bionic-fg quick menu state inside QuickMenu, not XServerScreen
Adding any locals to the XServerScreen composable makes D8 miscompile
it: the method has ~264 registers, and with the extra pressure D8
emitted an integer move (move/from16) for a CoroutineScope reference,
which ART's verifier rejects at class load (copy1 vX<-vY cat=1). Two
rounds of shaving locals moved the bad register (v270 -> v265) without
fixing it, so the feature is now register-neutral in XServerScreen:
QuickMenu already receives the container and creates BfgMenuState
itself, and the HUD fps multiplier reads the container extras directly
inside the fpsProvider lambda (a separate method).

Verified against the built APK: the faulting instruction is gone from
the dex and on-device dex2oat --compiler-filter=verify accepts the
class.
